+++ Plugin/CodeRebirth/src/Content/Items/Xui.cs
@@ -1,8 +1,7 @@
 using System.Collections;
 using System.Collections.Generic;
-using System.Linq;
-using CodeRebirth.src.Content.Unlockables;
 using CodeRebirth.src.Util;
+using CodeRebirthLib.ContentManagement.Enemies;
 using CodeRebirthLib.Extensions;
 using GameNetcodeStuff;
 using UnityEngine;
@@ -75,13 +74,13 @@ public class Xui : GrabbableObject
     {
         base.OnNetworkDespawn();
         
-        if(!CodeRebirthPlayerManager.Instance.IsServer) return;
+        if (!IsServer) return;
         
         int randomNumber = xuiRandom.Next(100);
         Plugin.ExtendedLogging($"Random Number: {randomNumber}");
         if (randomNumber <= 35)
         {
-            RoundManager.Instance.SpawnEnemyGameObject(this.transform.position, -1, -1, CodeRebirthUtils.EnemyTypes.Where(x => x.enemyName == "Masked").FirstOrDefault());
+            RoundManager.Instance.SpawnEnemyGameObject(this.transform.position, -1, -1, VanillaEnemies.MaskedPlayerEnemy);
             return;
         }
         List<PlayerControllerB> deadPlayers = new();
@@ -101,7 +100,7 @@ public class Xui : GrabbableObject
         }
         else if (randomNumber <= 70)
         {
-            RoundManager.Instance.SpawnEnemyGameObject(this.transform.position, -1, -1, CodeRebirthUtils.EnemyTypes.Where(x => x.enemyName == "Masked").FirstOrDefault());
+            RoundManager.Instance.SpawnEnemyGameObject(this.transform.position, -1, -1, VanillaEnemies.MaskedPlayerEnemy);
         }
     }
 }