+++ Plugin/CodeRebirth/src/Patches/StartOfRoundPatch.cs
@@ -38,7 +38,6 @@ static class StartOfRoundPatch
             if (CodeRebirthUtils.Instance == null)
             {
                 GameObject utilsInstance = GameObject.Instantiate(Plugin.Assets.UtilsPrefab);
-                utilsInstance.AddComponent<LightUpdateManager>();
                 SceneManager.MoveGameObjectToScene(utilsInstance, StartOfRound.Instance.gameObject.scene);
                 utilsInstance.GetComponent<NetworkObject>().Spawn();
                 Plugin.ExtendedLogging($"Created CodeRebirthUtils. Scene is: '{utilsInstance.scene.name}'");